requirements.txt
flask==2.0.1
azure-cognitiveservices-vision-computervision==0.9.0
python-dotenv==0.19.0
requests==2.26.0
Pillow==8.3.1
